cloud_function/dataset_upload_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.26.0
